# Improved: Paths Enumeration Performance