… tests/NexusMods.Paths.Tests/New/AbsolutePathTests/IOTests.cs