# About this PR

This PR optimises directory enumeration performance; to work in-line with our new pathing system introduced in https://github.com/Nexus-Mods/NexusMods.App/pull/137 , bringing massive memory efficiency improvements as well as improved fetching of FileSystem entries.

Implementation is based on some very little well known APIs in `System.IO.Enumeration` 😄, and improve over our current use of `System.IO.Directory` by going a tiny bit lower level.

# Benchmarks

The following benchmarks are performed on my Steam games folder; which consists of 217,039 files and 12,161 folders.  

```csharp
|                   Method |       Mean |    Error |   StdDev |       Gen0 |      Gen1 |      Gen2 | Allocated |
|------------------------- |-----------:|---------:|---------:|-----------:|----------:|----------:|----------:|
|       EnumerateFiles_New |   381.9 ms |  7.42 ms |  7.29 ms |  1000.0000 |         - |         - |  18.07 MB |
|       EnumerateFiles_Old |   393.6 ms |  5.74 ms |  5.37 ms |  6000.0000 | 2000.0000 |         - |  98.42 MB |
| EnumerateDirectories_New |   370.0 ms |  6.63 ms |  5.88 ms |          - |         - |         - |   5.83 MB |
| EnumerateDirectories_Old |   366.0 ms |  4.06 ms |  3.60 ms |          - |         - |         - |    7.3 MB |
| EnumerateFileEntries_New |   393.2 ms |  5.66 ms |  5.01 ms |  1000.0000 |         - |         - |  31.32 MB |
| EnumerateFileEntries_Old | 8,261.7 ms | 27.47 ms | 24.35 ms | 11000.0000 | 4000.0000 | 1000.0000 | 189.96 MB |
```

This benchmark is available as `EnumerateFiles.cs` (File Enumeration) in benchmark selector. Just change path to desired path as needed.

## The Benefits

### Massive Memory Savings.

All `AbsolutePath.Directory` entries are now deduplicated; which results in a massive saving in memory use.  

### Fast File Entry Fetching

Getting properties of the entries, e.g. Size, Last Modified is now an atomic operation and done as part of the search operation; therefore we are seeing massive acceleration on this end.

## Extra

A benchmark of my `C:\` drive.

```csharp
|                   Method |     Mean | Error |  StdDev |        Gen0 |       Gen1 |       Gen2 |  Allocated |
|------------------------- |---------:|------:|--------:|------------:|-----------:|-----------:|-----------:|
|       EnumerateFiles_New |  30.29 s |    NA | 0.045 s |  21000.0000 | 13000.0000 |  4000.0000 |  290.79 MB |
|       EnumerateFiles_Old |  30.44 s |    NA | 0.108 s |  67000.0000 | 23000.0000 | 12000.0000 |  961.82 MB |
| EnumerateDirectories_New |  30.35 s |    NA | 0.137 s |  13000.0000 | 11000.0000 |  3000.0000 |  185.15 MB |
| EnumerateDirectories_Old |  30.49 s |    NA | 0.006 s |  15000.0000 | 10000.0000 |  4000.0000 |  227.36 MB |
| EnumerateFileEntries_New |  30.63 s |    NA | 0.063 s |  29000.0000 | 15000.0000 |  5000.0000 |  399.15 MB |
| EnumerateFileEntries_Old | 172.37 s |    NA | 2.075 s | 127000.0000 | 40000.0000 | 18000.0000 | 1785.09 MB |
```